requirements.txt
Flask
Pillow
flask
keras
pymongo
tensorflow
